Add finance calculator task.

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 monthly_income = input("Enter your monthly income: ")
 total_expenses = input("Enter your total monthly expenses: ")
 
-monthly_savings = int(monthly_income) - int(total_expenses)
+monthly_savings = float(monthly_income) - float(total_expenses)
 print(f"Your monthly savings are: ${monthly_savings}")
 
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
